+++ hutool-core/src/main/java/cn/hutool/core/lang/SimpleCache.java
@@ -1,13 +1,11 @@
 package cn.hutool.core.lang;
 
+import cn.hutool.core.lang.func.Func0;
+
 import java.io.Serializable;
 import java.util.Map;
 import java.util.WeakHashMap;
-import java.util.concurrent.locks.ReentrantReadWriteLock;
-import java.util.concurrent.locks.ReentrantReadWriteLock.ReadLock;
-import java.util.concurrent.locks.ReentrantReadWriteLock.WriteLock;
-
-import cn.hutool.core.lang.func.Func0;
+import java.util.concurrent.locks.StampedLock;
 
 /**
  * 简单缓存，无超时实现，使用{@link WeakHashMap}实现缓存自动清理
@@ -21,10 +19,9 @@ public class SimpleCache<K, V> implements Serializable{
 	
 	/** 池 */
 	private final Map<K, V> cache = new WeakHashMap<>();
-	
-	private final ReentrantReadWriteLock cacheLock = new ReentrantReadWriteLock();
-	private final ReadLock readLock = cacheLock.readLock();
-	private final WriteLock writeLock = cacheLock.writeLock();
+
+	// 乐观读写锁
+	private final StampedLock lock = new StampedLock ();
 
 	/**
 	 * 从缓存池中查找值
@@ -33,15 +30,12 @@ public class SimpleCache<K, V> implements Serializable{
 	 * @return 值
 	 */
 	public V get(K key) {
-		// 尝试读取缓存
-		readLock.lock();
-		V value;
+		long stamp = lock.readLock();
 		try {
-			value = cache.get(key);
+			return cache.get(key);
 		} finally {
-			readLock.unlock();
+			lock.unlockRead(stamp);
 		}
-		return value;
 	}
 	
 	/**
@@ -52,12 +46,25 @@ public class SimpleCache<K, V> implements Serializable{
 	 * @return 值对象
 	 */
 	public V get(K key, Func0<V> supplier) {
-		V v = get(key);
-		if (null == v && null != supplier) {
-			writeLock.lock();
-			try {
-				// 双重检查锁
+		if(null == supplier){
+			return get(key);
+		}
+
+		long stamp = lock.readLock();
+		V v;
+		try{
+			v = cache.get(key);
+			if (null == v) {
+				// 尝试转换独占写锁
+				long writeStamp = lock.tryConvertToWriteLock(stamp);
+				if(0 == writeStamp){
+					// 转换失败，手动更新为写锁
+					lock.unlockRead(stamp);
+					writeStamp = lock.writeLock();
+				}
+				stamp = writeStamp;
 				v = cache.get(key);
+				// 双重检查，防止在竞争锁的过程中已经有其它线程写入
 				if(null == v) {
 					try {
 						v = supplier.call();
@@ -66,9 +73,9 @@ public class SimpleCache<K, V> implements Serializable{
 					}
 					cache.put(key, v);
 				}
-			} finally {
-				writeLock.unlock();
 			}
+		} finally {
+			lock.unlock(stamp);
 		}
 		return v;
 	}
@@ -80,11 +87,12 @@ public class SimpleCache<K, V> implements Serializable{
 	 * @return 值
 	 */
 	public V put(K key, V value){
-		writeLock.lock();
+		// 独占写锁
+		final long stamp = lock.writeLock();
 		try {
 			cache.put(key, value);
 		} finally {
-			writeLock.unlock();
+			lock.unlockWrite(stamp);
 		}
 		return value;
 	}
@@ -96,11 +104,12 @@ public class SimpleCache<K, V> implements Serializable{
 	 * @return 移除的值
 	 */
 	public V remove(K key) {
-		writeLock.lock();
+		// 独占写锁
+		final long stamp = lock.writeLock();
 		try {
 			return cache.remove(key);
 		} finally {
-			writeLock.unlock();
+			lock.unlockWrite(stamp);
 		}
 	}
 
@@ -108,11 +117,12 @@ public class SimpleCache<K, V> implements Serializable{
 	 * 清空缓存池
 	 */
 	public void clear() {
-		writeLock.lock();
+		// 独占写锁
+		final long stamp = lock.writeLock();
 		try {
 			this.cache.clear();
 		} finally {
-			writeLock.unlock();
+			lock.unlockWrite(stamp);
 		}
 	}
 }